Add creation of Jira WBS groups and better curl return code
Add creation of the Jira WBS groups, including the nesting of the COU
active group and adding the Jira Group Name identifier. Also add
handling of the curl return code for the query to the XCDB API so that
if anything other than a 200 is returned log error and go on to the next
CO Person record.